+++ src/bp/core/events/event-engine.ts
@@ -1,8 +1,8 @@
 import * as sdk from 'botpress/sdk'
+import { Queue } from 'core/events/queue'
 import { TimedPerfCounter } from 'core/misc/timed-perf'
 import { WellKnownFlags } from 'core/sdk/enums'
 import { incrementMetric } from 'core/services/monitoring'
-import { Queue } from 'core/services/queue'
 import { inject, injectable, tagged } from 'inversify'
 import joi from 'joi'
 import _ from 'lodash'
