+++ src/bp/core/events/event-repository.ts
@@ -1,8 +1,8 @@
 import * as sdk from 'botpress/sdk'
+import Database from 'core/database'
 import { inject, injectable } from 'inversify'
 import _ from 'lodash'
 
-import Database from '../database'
 import { TYPES } from '../types'
 
 export interface EventRepository {
