+++ src/bp/core/services/services.inversify.ts
@@ -1,6 +1,8 @@
 import { IO } from 'botpress/sdk'
 import LicensingService from 'common/licensing-service'
 import { EventEngine } from 'core/events/event-engine'
+import { Queue } from 'core/events/queue'
+import MemoryQueue from 'core/events/queue/memory-queue'
 import { DialogContainerModule } from 'core/services/dialog/dialog.inversify'
 import { CEJobService, JobService } from 'core/services/job-service'
 import { ContainerModule, interfaces } from 'inversify'
@@ -28,8 +30,6 @@ import { MediaServiceProvider } from './media'
 import { CEMonitoringService, MonitoringService } from './monitoring'
 import { NLUService } from './nlu/nlu-service'
 import { NotificationsService } from './notification/service'
-import { Queue } from './queue'
-import MemoryQueue from './queue/memory-queue'
 import RealtimeService from './realtime'
 import { StatsService } from './stats-service'
